# api_specification.yaml
openapi: 3.0.1
...
info:
  title: Webdp API - Reworked
  version: "2.0"
  description: "Welcome to the official OpenAPI documentation for WebDP, our versatile\
    \ API designed to provide transparent interoperability with a range of differentially\
    \ private frameworks."
  license:
    name: Mozilla Public License version 2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/

# -----------------------------------
servers:
- url: //localhost:8000/

# -----------------------------------
security:
  - BearerTokenAuth: [] # Throughout the API

# -----------------------------------
paths:
...
  /v1/datasets:
    get:
      tags:
      - datasets
      summary: Gets all datasets which requester has access to.
      description: "Requester needs role admin or curator, or needs granted access\
        \ via budget allocation."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entity.DatasetInfo'
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
      - datasets
      summary: Creates a dataset.
      description: Requester needs role curator. New owner of dataset needs role curator.
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.DatasetCreate'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.Id'
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
  /v1/dataset/{datasetId}:
    get:
      tags:
      - datasets
      summary: Gets all datasets which requester has access to.
      description: "Requester needs role admin or curator, or needs granted access\
        \ via budget allocation."
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.DatasetInfo'
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags:
      - datasets
      summary: Delete a dataset.
      description: Requester needs to be the owner of the dataset.
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags:
      - datasets
      summary: Update a dataset.
      description: |-
        Update name, owner or total budget of a dataset.
        Requester needs to be curator and owner of the dataset.
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.DatasetCreate'
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
  /v1/dataset/{datasetId}/upload:
    post:
      tags:
      - datasets
      summary: Upload a dataset.
      description: Requester needs to be the owner of the dataset.
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: CSV Data
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: csvData
  /v2/datasets:
    get:
      tags:
      - datasets
      summary: Gets all datasets which requester has access to.
      description: "Requester needs role admin or curator, or needs granted access\
        \ via budget allocation."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entity.DatasetInfo'
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
      - datasets
      summary: Creates a dataset.
      description: Requester needs role curator. New owner of dataset needs role curator.
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.DatasetCreate'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.Id'
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
  /v2/datasets/{datasetId}:
    get:
      tags:
      - datasets
      summary: Gets all datasets which requester has access to.
      description: "Requester needs role admin or curator, or needs granted access\
        \ via budget allocation."
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.DatasetInfo'
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags:
      - datasets
      summary: Delete a dataset.
      description: Requester needs to be the owner of the dataset.
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags:
      - datasets
      summary: Update a dataset.
      description: |-
        Update name, owner or total budget of a dataset.
        Requester needs to be curator and owner of the dataset.
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.DatasetCreate'
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
  /v2/datasets/{datasetId}/upload:
    post:
      tags:
      - datasets
      summary: Upload a dataset.
      description: Requester needs to be the owner of the dataset.
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: CSV Data
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: csvData

  # Budgets
...
components:
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/response.Error"

  # Schemas
  schemas:
...
    entity.Budget:
      type: object
      properties:
        delta:
          type: number
        epsilon:
          type: number
    entity.ColumnSchema:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/entity.DataType'
    entity.DataType:
      type: object
      properties:
        type:
          type: object
...
    entity.DatasetCreate:
      type: object
      properties:
        name:
          type: string
        owner:
          type: string
        privacy_notion:
          type: string
        schema:
          type: array
          items:
            $ref: '#/components/schemas/entity.ColumnSchema'
        total_budget:
          $ref: '#/components/schemas/entity.Budget'
    entity.DatasetInfo:
      type: object
      properties:
        created_time:
          type: string
        id:
          type: integer
        loaded:
          type: boolean
        loaded_time:
          type: string
        name:
          type: string
        owner:
          type: string
        privacy_notion:
          type: string
        schema:
          type: array
          items:
            $ref: '#/components/schemas/entity.ColumnSchema'
        total_budget:
          $ref: '#/components/schemas/entity.Budget'
        updated_time:
          type: string
...
    response.Error:
      description: An error message
      type: object
      properties:
        detail:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
...
    response.Id:
      type: object
      properties:
        id:
          type: integer
  securitySchemes:
    BearerTokenAuth:
      type: http
      scheme: basic